+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.3.5 - TBD
+
+- Add helper methods to create and verify detached signatures:
+    - Add `CoseSignBuilder` methods `add_detached_signature` and `try_add_detached_signature`.
+    - Add `CoseSign` method `verify_detached_signature`.
+    - Add `CoseSign1Builder` methods `create_detached_signature` and `try_create_detached_signature`.
+    - Add `CoseSign1` method `verify_detached_signature`.
+
 ## 0.3.4 - 2023-01-25
 
 - Add non-default `std` feature that turns on `impl Error for CoseError`.